refactor(model-components): update import paths and types
- Change import path for AvailableModel in ModelSelector
- Update models type in ModelCatalog from ModelCatalogEntry to Model
- Refactor imports in useModels for better organization
- Adjust import of ModelStatus in ModelsScreen
